Unwrap PowerShell commands for exec policy
