+++ codex-rs/shell-command/src/command_safety/mod.rs
@@ -3,3 +3,6 @@ mod powershell_parser;
 pub mod is_dangerous_command;
 pub mod is_safe_command;
 pub(crate) mod windows_safe_commands;
+
+pub(crate) use powershell_parser::PowershellParseOutcome;
+pub(crate) use powershell_parser::parse_with_powershell_ast;
